#137 refactor: extract PositionBreakdownCard + RecommendedTradeCard from TeamAnalyzerView
